+++ indico/web/client/js/custom_elements/TipBase.js
@@ -71,7 +71,8 @@ export class TipBase extends CustomElementBase {
     this.addEventListener('x-connect', () => {
       this.contentMutationObserver = new MutationObserver(() => {
         this.updatePosition();
-      }).observe(this.$tip, {
+      });
+      this.contentMutationObserver.observe(this.$tip, {
         subtree: true,
         childList: true,
         characterData: true,