+++ indico/web/client/js/react/components/category_navigator/context.js
@@ -1,0 +1,20 @@
+// This file is part of Indico.
+// Copyright (C) 2002 - 2026 CERN
+//
+// Indico is free software; you can redistribute it and/or
+// modify it under the terms of the MIT License; see the
+// LICENSE file for more details.
+
+import {createContext, useContext} from 'react';
+
+const CategoryNavigatorContext = createContext(null);
+
+export function useCategoryNavigator() {
+  const context = useContext(CategoryNavigatorContext);
+  if (!context) {
+    throw new Error('useCategoryNavigator must be used within CategoryNavigatorProvider');
+  }
+  return context;
+}
+
+export default CategoryNavigatorContext;